Menu系のコンポーネントをまとめ、型を別ファイルに分ける
