+++ src/components/Menu/MenuBar/TitleBarEditorSwitcher.vue
