+++ src/components/Menu/MenuButton.vue
@@ -43,8 +43,8 @@
 <script setup lang="ts">
 import { computed, ref, watch } from "vue";
 import { debounce } from "quasar";
-import MenuItem from "@/components/MenuItem.vue";
-import { MenuItemData } from "@/components/BaseMenuBar.vue";
+import MenuItem from "./MenuItem.vue";
+import { MenuItemData } from "./type";
 import { useStore } from "@/store";
 
 const props = withDefaults(
