+++ src/components/Menu/MenuItem.vue
@@ -72,7 +72,7 @@
 
 <script setup lang="ts">
 import { ref, computed, watch } from "vue";
-import type { MenuItemData } from "@/components/BaseMenuBar.vue";
+import { MenuItemData } from "./type";
 import { useStore } from "@/store";
 import { hotkeyActionSchema } from "@/type/preload";
 const props = withDefaults(
