+++ src/components/Talk/AccentPhrase.vue
@@ -209,8 +209,8 @@
 import { computed, ref } from "vue";
 import AudioAccent from "./AudioAccent.vue";
 import AudioParameter from "./AudioParameter.vue";
-import { MenuItemButton } from "@/components/BaseMenuBar.vue";
-import ContextMenu from "@/components/ContextMenu.vue";
+import { MenuItemButton } from "@/components/Menu/type";
+import ContextMenu from "@/components/Menu/ContextMenu.vue";
 import { useStore } from "@/store";
 import { AudioKey, MoraDataType } from "@/type/preload";
 import { Mora } from "@/openapi/models/Mora";
