+++ src/components/Talk/AudioCell.vue
@@ -113,11 +113,8 @@
 import { computed, watch, ref, nextTick } from "vue";
 import { QInput } from "quasar";
 import CharacterButton from "@/components/CharacterButton.vue";
-import {
-  MenuItemButton,
-  MenuItemSeparator,
-} from "@/components/BaseMenuBar.vue";
-import ContextMenu from "@/components/ContextMenu.vue";
+import { MenuItemButton, MenuItemSeparator } from "@/components/Menu/type";
+import ContextMenu from "@/components/Menu/ContextMenu.vue";
 import { useStore } from "@/store";
 import { AudioKey, SplitTextWhenPasteType, Voice } from "@/type/preload";
 import { SelectionHelperForQInput } from "@/helpers/SelectionHelperForQInput";
